Flytte metoder fra EksporteringService til EksporteringSErviceUtils slik at de kan brukes i EksporteringPerStatistikkKategoriService.
