+++ src/test/java/no/nav/arbeidsgiver/sykefravarsstatistikk/api/eksportering/autoeksport/EksporteringServiceMockTest.java
@@ -98,146 +98,6 @@ public class EksporteringServiceMockTest {
         true);
   }
 
-  @Test
-  public void getListeAvVirksomhetEksportPerKvartal_tar_hensyn_til_begrensning_i_eksportering_også_med_emptyList() {
-    when(eksporteringRepository.hentVirksomhetEksportPerKvartal(__2020_2)).thenReturn(
-        Collections.emptyList());
-
-    List<VirksomhetEksportPerKvartal> emptyList1 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().utenBegrensning(),
-            eksporteringRepository
-        );
-    List<VirksomhetEksportPerKvartal> emptyList2 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().medBegrensning(100),
-            eksporteringRepository
-        );
-
-    assertThat(emptyList1.size()).isEqualTo(0);
-    assertThat(emptyList2.size()).isEqualTo(0);
-  }
-
-  @Test
-  public void getListeAvVirksomhetEksportPerKvartal_tar_hensyn_til_begrensning_i_eksportering() {
-    when(eksporteringRepository.hentVirksomhetEksportPerKvartal(__2020_2))
-        .thenReturn(
-            Arrays.asList(
-                new VirksomhetEksportPerKvartal(
-                    new Orgnr("987654321"),
-                    __2020_2,
-                    false
-                ),
-                new VirksomhetEksportPerKvartal(
-                    new Orgnr("999999999"),
-                    __2020_2,
-                    false
-                ),
-                new VirksomhetEksportPerKvartal(
-                    new Orgnr("888888888"),
-                    __2020_2,
-                    false
-                )));
-
-    List<VirksomhetEksportPerKvartal> list1 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().utenBegrensning(),
-            eksporteringRepository
-        );
-    List<VirksomhetEksportPerKvartal> list2 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().medBegrensning(2),
-            eksporteringRepository
-        );
-    List<VirksomhetEksportPerKvartal> list3 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().medBegrensning(3),
-            eksporteringRepository
-        );
-    List<VirksomhetEksportPerKvartal> list4 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().medBegrensning(100),
-            eksporteringRepository
-        );
-
-    assertThat(list1.size()).isEqualTo(3);
-    assertThat(list2.size()).isEqualTo(2);
-    assertThat(list3.size()).isEqualTo(3);
-    assertThat(list4.size()).isEqualTo(3);
-  }
-
-  @Test
-  public void getListeAvVirksomhetEksportPerKvartal_tar_hensyn_til_begrensning_i_eksportering_og_eksportert_flag__med_mye_data() {
-    when(eksporteringRepository.hentVirksomhetEksportPerKvartal(__2020_2))
-        .thenReturn(bigList(90000, 430000));
-
-    List<VirksomhetEksportPerKvartal> ikkeBegrenset =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().utenBegrensning(),
-            eksporteringRepository
-        );
-    assertThat(ikkeBegrenset.size()).isEqualTo(430000);
-
-    List<VirksomhetEksportPerKvartal> begrensetTil10 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().medBegrensning(10),
-            eksporteringRepository
-        );
-    assertThat(begrensetTil10.size()).isEqualTo(10);
-  }
-
-  @Test
-  public void getListeAvVirksomhetEksportPerKvartal_tar_hensyn_til_begrensning_i_eksportering_og_eksportert_flag() {
-    when(eksporteringRepository.hentVirksomhetEksportPerKvartal(__2020_2))
-        .thenReturn(
-            Arrays.asList(
-                new VirksomhetEksportPerKvartal(
-                    new Orgnr("987654321"),
-                    __2020_2,
-                    false
-                ),
-                new VirksomhetEksportPerKvartal(
-                    new Orgnr("999999999"),
-                    __2020_2,
-                    false
-                ),
-                new VirksomhetEksportPerKvartal(
-                    new Orgnr("888888888"),
-                    __2020_2,
-                    true
-                )));
-
-    List<VirksomhetEksportPerKvartal> ikkeBegrenset =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().utenBegrensning(),
-            eksporteringRepository
-        );
-    List<VirksomhetEksportPerKvartal> begrensetMed1 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().medBegrensning(1),
-            eksporteringRepository
-        );
-    List<VirksomhetEksportPerKvartal> begrensetMed100 =
-        service.getListeAvVirksomhetEksportPerKvartal(
-            __2020_2,
-            EksporteringBegrensning.build().medBegrensning(100),
-            eksporteringRepository
-        );
-
-    assertThat(ikkeBegrenset.size()).isEqualTo(2);
-    assertThat(begrensetMed1.size()).isEqualTo(1);
-    assertThat(begrensetMed100.size()).isEqualTo(2);
-  }
 
   @Test
   public void eksporter_returnerer_antall_rader_eksportert() {
